Fixed some issues from PR review
Modem design image

modified diagram

renamed diagram

Create design.md

Update design.md

Updated diagram

Update design.md

switched diagram back to png

Update design.md

Update design.md

Untested initial command with tests

Added gmock/gtest submodules

Revert "Added gmock/gtest submodules"

This reverts commit b975f3be8bfe28b92b6ebb31018bdae8b7857997.

found out googlemock and googletest are now merged, only need one.

Added googletest submodule

import example make file

Added *Ptr methods to firebase, get test mostly compiling

Broke FirebaseHttpClient out from Firebase to allow greater portability and testability.

re check in commands and add input stream

Working local command test

Added base test for set

Added base test for set (actually this time)

Cleanup set test and add error test

Clean up get test and added failure handling

Added remove command and test

Added push command and test

Minor cleanup to set test

Added begin command and test

Fixed all tests operating on the assumption that readLine() included the \r\n

Add base for SerialTransciever.

Getting serial example together

Cleaning up

Changes to make both arduino and tests compile

All existing examples work now.

changes to get SerialHost building

Fixed up Stream example

Fixes to get FirebaseSerialHost.ino working!

Switch to huzzah module.

All examples are working

unit tests are all working with new changes

Added stream command using simple syntax.

Fixes from testing with serial modem

Reorganized the library to comply with arduino 1.5+ library standards https://github.com/arduino/Arduino/wiki/Arduino-IDE-1.5:-Library-specification

Fixed some issues from PR review

Modem design image

modified diagram

renamed diagram

Create design.md

Update design.md

Updated diagram

Update design.md

switched diagram back to png

Update design.md

Update design.md

Untested initial command with tests

Added gmock/gtest submodules

Revert "Added gmock/gtest submodules"

This reverts commit b975f3be8bfe28b92b6ebb31018bdae8b7857997.

found out googlemock and googletest are now merged, only need one.

Added googletest submodule

import example make file

Added *Ptr methods to firebase, get test mostly compiling

Broke FirebaseHttpClient out from Firebase to allow greater portability and testability.

re check in commands and add input stream

Working local command test

Added base test for set

Added base test for set (actually this time)

Cleanup set test and add error test

Clean up get test and added failure handling

Added remove command and test

Added push command and test

Minor cleanup to set test

Added begin command and test

Fixed all tests operating on the assumption that readLine() included the \r\n

Add base for SerialTransciever.

Getting serial example together

Cleaning up

Changes to make both arduino and tests compile

All existing examples work now.

changes to get SerialHost building

Fixed up Stream example

Fixed up Stream example

Fixes to get FirebaseSerialHost.ino working!

Switch to huzzah module.

All examples are working

unit tests are all working with new changes

Add modified submodule

Added stream command using simple syntax.

Fixes from testing with serial modem

Reorganized the library to comply with arduino 1.5+ library standards https://github.com/arduino/Arduino/wiki/Arduino-IDE-1.5:-Library-specification

Fixed some issues from PR review

re check in commands and add input stream

Working local command test

Update design.md

Update design.md

Added base test for set

Added base test for set (actually this time)

Cleanup set test and add error test

Added remove command and test

Added push command and test

Minor cleanup to set test

Added begin command and test

Fixed all tests operating on the assumption that readLine() included the \r\n

Add base for SerialTransciever.

Getting serial example together

Cleaning up

Changes to make both arduino and tests compile

All existing examples work now.

changes to get SerialHost building

Fixed up Stream example

Fixed up Stream example

Fixes to get FirebaseSerialHost.ino working!

Switch to huzzah module.

All examples are working

unit tests are all working with new changes

Add modified submodule

Added stream command using simple syntax.

Fixes from testing with serial modem

Reorganized the library to comply with arduino 1.5+ library standards https://github.com/arduino/Arduino/wiki/Arduino-IDE-1.5:-Library-specification

Fixed some issues from PR review
