+++ src/Firebase.h
@@ -40,6 +40,10 @@ class Firebase {
 
   Firebase(const Firebase&) = delete;
 
+  Firebase(const Firebase&) = delete;
+
+  Firebase(const Firebase&) = delete;
+
   // Fetch auth string back.
   const String& auth();
 
@@ -191,6 +195,6 @@ class FirebaseStream : public FirebaseCall {
 
  private:
   FirebaseError _error;
- };
+};
 
 #endif // firebase_h
